# packaging/npm/token-free-gateway/agentmesh/services/api/openapi.yaml
openapi: 3.1.0
info:
  title: MuhanAI AgentMesh API
  version: 0.1.0
  description: |
    HTTP contract for the @agentmesh/api service (Fastify). Routes are split across
    the following files under `services/api/src/`:
      - agents-routes.ts
      - network-routes.ts
      - knowledge-routes.ts
      - compute-routes.ts
      - hivebear-routes.ts
      - semantic-routes.ts
      - noema-routes.ts

    Body validation is performed with Zod schemas — request shapes below mirror
    the corresponding Zod definitions. Error responses use `{ error: string }`
    unless otherwise noted.
  license:
    name: MuhanAI Token-Free Gateway
servers:
  - url: http://localhost:8787
    description: Local Fastify dev server (default port)
paths:
...
  /api/compute/tribute:
    post:
      tags: [compute]
      summary: Submit a tribute task to the compute mesh
      operationId: submitTribute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TributeSubmitRequest"
      responses:
        "202":
          description: Queued ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TributeTicket"
        "400":
          $ref: "#/components/responses/ZodError"
...
  /api/compute/tribute/next:
    post:
      tags: [compute]
      summary: Pop the next runnable tribute
      operationId: nextTribute
      responses:
        "200":
          description: Next ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TributeTicket"
        "404":
          $ref: "#/components/responses/ErrorString"
  /api/compute/tribute/{taskId}:
    delete:
      tags: [compute]
      summary: Cancel and remove a tribute
      operationId: removeTribute
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Removed
        "404":
          $ref: "#/components/responses/ErrorString"
  /api/compute/projects/weight:
    post:
      tags: [compute]
      summary: Set a project's weight in the compute router
      operationId: setProjectWeight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, weight]
              properties:
                projectId: { type: string }
                weight: { type: number }
      responses:
        "200":
          description: Set
        "400":
          $ref: "#/components/responses/ErrorString"

  # ─── hivebear ──────────────────────────────────────────────────────────
...
components:
  responses:
    ZodError:
      description: Zod validation failure
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    ErrorString:
      description: Plain string error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
...
    TributeSubmitRequest:
      type: object
      required: [taskId]
      properties:
        taskId: { type: string, minLength: 1 }
        projectId: { type: string }
        networkShare: { type: number, minimum: 0, maximum: 1, default: 0.5 }
        personalShare: { type: number, minimum: 0, maximum: 1, default: 0.5 }
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
    TributeTicket:
      allOf:
        - $ref: "#/components/schemas/TributeSubmitRequest"
        - type: object
          properties:
            status:
              type: string
              enum: [queued, running, done, failed]
...